Adds The Match Making Algorithm
This commit adds in the logic to match up teams based on their ELO. The method I chose is the simulated annealing approach. Essentially, we randomly assign each guardian to a team. Then we start iterating and randomly swapping two guardians on two different teams. Then, we measure the ELO difference of those two teams from before and after the swap and take the difference. If it is less than zero, that means the swap is favorable and we are closer to having equal ELOs. Otherwise, we roll a random probability and compare it to the change in ELO. If the rolled probability is lower, then we will still keep the assignment and keep moving. This allows for randomness in the way the teams settle to let it reach a more equal state