Add frontend views and functionality (#6)
* Add in Base Models

This commit adds in the base models that will be used for the web app

* Add Average Eff To Guardian

This commit adds the average efficiency field to the guardian object because i think this will be a field we will want to track over time

* Averages Usually Arent Ints

This commit changes the average efficiency field from an int to double because why would that ever be an int

* Break Out Model Concerns

This commit breaks out different guardian conerns from the guardian class.

A guardians elo will now be represented by their own table so we can better track elo changes over times.
A guardians efficiency will also now be tracked by its own table for the same reasons.

This commit also changes the structure of the match object so it is not limited by two teams.

* Add Models Folder Back to csproj

* Add In Base CRUD For Guardians

This commit adds in the base file structure and code to be able to get all guardians, get a specific guardian, create a guardian, and also delete them.

This commit also adds in the logic to update a guardians total kills and total deaths

* Remove Old Migration Logic

This commit removes leftover code from how I was thinking about handling migrations before.

Now we will just use entity framework to handle all that jazz

* Add Get Single Guardian Endpoint

This commit adds in the logix to get a single guardian from the DB.

* Update DB Models and Migrations

* Add in the updated DB context object. Originally we were planning on using multiple context objects for our models but that seems like too much overhead. It is now simplified to use a single context

* Add ID Prefixes as static classes to make sure they remain consistent and if we need to change them we only have to change them in a single place

* Add Front To Back Controls For Matches

This commit adds the front to back workflow for adding matches, updating matches, deleting matches, and all that goodness. This commit also adds in the workflow for updating the postgres database on startup

* Remove Unused Using Directives

* Delete Old Migration Files

This commit removes some broken migration files. I was using a database context class that is no longer used. It should get deleted but these migration files and snapshots depend on it

* Add EF Core Dependencies

* Add Initial Migrations Files

* Add Dotnet EF Tool To API Package

* Downgrade EF CLI Version

* Add Connection URI Parser

Turns out, Heroku provides the connection to the Postgres db as a URI. C# needs it to be a connection string. This commit adds in a little bit of parsing to get out the necessary information into a form we can use

* Add Quick Nav Links

This commit adds in nav links so that we can navigate through the UI to each of the pages.

* Add In Formatting to Efficiency

This commit adds in some formatting to the efficiency on the main page, and also adds nav back to the home page from the all guardians page

* Add In Auth Middleware

This commit adds in middleware to the API that whitelists IPs so we can block requests from anyone whos not on the Admin list

* Remove Access to All Endpoints

* Add My IP To WHitelist

* Remove IP Whitelist from Source

* Change Where Looking Admin Safe List Env

* Add Debug Logging Statements To Middleware Cos Its No Working

* Rework IP Logic Cos It STILL No Work :(

* Still Trying :(

* More Logging :(

* Change Up the Logic Lil Bit

* Try More Logix CoS ITS STILL NOT WORKING

* ZOMG It Works

This commit removes some of the debug log statements that aren't necessary anymore

* Remove Auth Guards

* FaceLift With Styles

* Remove Client Secret From Source

* Add IN Auth Guards To All Non Get Routes

* More FaceLifty Types of Things

* Fix Styling Bug And Add Minified CSS

* Last Commit On This PR

This commit adds in Guardian Performance tracking over time via graphs and charts and is hella dope

* Yep Forgot To commit A Buncho Stuff. WHoops

* Fix Divide by Zero Bug

* Update Bundler

* Update BundleConfig

* Fix Production Layer Script Importing

* Add Icons

* Actually Commit Files

* Add Delete Button For Matches

* Add Delete Guardian Functionality to UI

* Fix Guardian Display Bug and Table Overlap Styles

* Fix BadBadBug

This commit fixes a bug that would set efficiency to zero if a guardian went flawless in a match. That would be very much no bueno

* Add Sorting To Guardians Grid

* Fix Row Ordering

* Rename Guardian Controller to Conform to REST

* Update Nav To Reflect Controller Change

* Change the Delete Method from POST to DELETE

'tis more RESTful indeed

* Turns Out I Never Commited Bootstrap Styles. Toit

* Frontend Now Makes Delete Request Instead of Post to Delete Matches